Drop GuiceMainSupport; merge into MainSupport
