+++ gshell-core/src/main/java/com/planet57/gshell/help/HelpPageManagerImpl.java
@@ -32,7 +32,7 @@ import com.planet57.gshell.alias.NoSuchAliasException;
 import com.planet57.gshell.command.resolver.CommandResolver;
 import com.planet57.gshell.command.resolver.Node;
 import com.planet57.gshell.event.EventManager;
-import com.planet57.gshell.guice.BeanContainer;
+import com.planet57.gshell.internal.BeanContainer;
 import org.sonatype.goodies.common.ComponentSupport;
 
 import static com.google.common.base.Preconditions.checkNotNull;
