+++ gshell-dist/gshell-personality/src/main/java/com/planet57/gshell/Main.java
@@ -19,7 +19,6 @@ import java.util.List;
 
 import com.google.inject.Module;
 import com.planet57.gshell.branding.Branding;
-import com.planet57.gshell.guice.GuiceMainSupport;
 import com.planet57.gshell.logging.LoggingSystem;
 import com.planet57.gshell.logging.logback.LogbackLoggingSystem;
 
@@ -32,7 +31,7 @@ import javax.annotation.Nonnull;
  * @since 2.0
  */
 public class Main
-    extends GuiceMainSupport
+    extends MainSupport
 {
   @Override
   protected Branding createBranding() {
