+++ gshell-maven-plugin/src/main/java/com/planet57/gshell/maven/RunMojo.java
@@ -24,7 +24,7 @@ import com.planet57.gshell.branding.BrandingSupport;
 import com.planet57.gshell.branding.License;
 import com.planet57.gshell.command.IO;
 import com.planet57.gshell.execute.ExitNotification;
-import com.planet57.gshell.guice.BeanContainer;
+import com.planet57.gshell.internal.BeanContainer;
 import com.planet57.gshell.logging.LoggingSystem;
 import com.planet57.gshell.shell.ShellImpl;
 import com.planet57.gshell.util.io.PrintBuffer;
